Add `:init_script` option to `Snex.Interpreter`
A init script is a Python code snippet that will run when
the interpreter is started and before it runs any commands.
Failing to run the script will cause the process initialization
to error.  The variable context left by the script will become
the initial context for all `Snex.make_env/3` calls using this
interpreter.